+++ src/shared/components/Header/Header.module.scss
@@ -4,62 +4,10 @@
   top: 0px;
   left: 0px;
   right: 0px;
-}
-
-.top {
   display: flex;
   align-items: center;
   padding: 10px 10px;
   justify-content: space-between;
   background: rgb(53, 53, 53);
   border-bottom: 1px solid rgb(66, 66, 66);
-  width: 100%;
-}
-
-.navbar {
-  display: flex;
-  gap: 20px;
-  align-items: center;
-}
-.link {
-  text-decoration: none;
-  font-size: 16px;
-  font-weight: 600;
-  position: relative;
-  color: #fff;
-  &::after {
-    position: absolute;
-    content: " ";
-    width: 0%;
-    bottom: -2px;
-    left: 0px;
-    height: 1px;
-    background: white;
-    transition: width 300ms ease-out;
-  }
-  &:hover {
-    &::after {
-      width: 100%;
-    }
-  }
-}
-
-.unAuthBtn {
-  padding: 10px 20px;
-  border-radius: 10px;
-  background: transparent;
-  border: 1px solid rgb(82, 82, 82);
-  font-weight: 600;
-  font-size: 16px;
-  cursor: pointer;
-  transition: all 300ms ease-out;
-  &:hover {
-    background: rgb(58, 58, 58);
-  }
-}
-
-.logInLinks {
-  display: flex;
-  gap: 20px;
-  margin-right: 10px;
 }
