+++ src/shared/components/Header/components/Controls/Controls.module.scss
@@ -1,0 +1,28 @@
+.unAuthBtn {
+  padding: 10px 20px;
+  border-radius: 10px;
+  background: transparent;
+  border: 1px solid rgb(82, 82, 82);
+  font-weight: 600;
+  font-size: 16px;
+  cursor: pointer;
+  transition: all 300ms ease-out;
+  &:hover {
+    background: rgb(58, 58, 58);
+  }
+}
+
+.logInLinks {
+  display: flex;
+  gap: 20px;
+  margin-right: 10px;
+}
+
+.controlsWrapper {
+  opacity: 1;
+  transition: opacity 300ms ease-out;
+}
+
+.hidden {
+  opacity: 0;
+}
