+++ src/shared/components/Header/components/Navbar/Navbar.module.scss
@@ -1,0 +1,27 @@
+.navbar {
+  display: flex;
+  gap: 20px;
+  align-items: center;
+}
+.link {
+  text-decoration: none;
+  font-size: 16px;
+  font-weight: 600;
+  position: relative;
+  color: #fff;
+  &::after {
+    position: absolute;
+    content: " ";
+    width: 0%;
+    bottom: -2px;
+    left: 0px;
+    height: 1px;
+    background: white;
+    transition: width 300ms ease-out;
+  }
+  &:hover {
+    &::after {
+      width: 100%;
+    }
+  }
+}
