--- src/shared/firebase/actions/logOut.tsx
@@ -1,7 +1,0 @@
-import { Auth, signOut } from "firebase/auth";
-
-export const LogOut = async (auth: Auth) => {
-  await signOut(auth).catch((error) => {
-    console.log(error);
-  });
-};
